Move `DetailsViewController` to production
