--- CEPSearch/CEPiOS/CEPViewController.swift
@@ -1,8 +1,0 @@
-//
-//  CEPViewController.swift
-//  CEPiOS
-//
-//  Created by Ivo on 28/01/24.
-//
-
-import Foundation
